+++ src/events/guild-delete.js
@@ -3,8 +3,8 @@ const Logger = require('../utils/logger');
 
 module.exports = class GuildCreateEvent extends SebediusEvent {
   name = 'guildDelete';
-  async execute(client, guild) {
-    const total = client.guilds.cache.size;
+  async execute(guild) {
+    const total = this.bot.guilds.cache.size;
     const msg = `🚪 Guild #${total} | Left: ${guild.name} (${guild.id})`;
     Logger.event(msg);
     await this.bot.database.grabGuild(guild, {